Fix polygon generation
This fix is intended to create a polygon that matches with its preview in one drawing step (needed when we want to draw with Contour or Polygon tool with custom brushes with alpha content). Before this fix, the polygon was being created in a first step, and then a second step that patches the contour (over writing the Image with an extra joinStroke execution of the entire contour).

- Added createUnion function in polygon.cpp to force drawing of the input points in each scan line render.
- Added algo_line_continuous inside polygon function to interpolate holes between input vertices.
- Deleted extra joinStroke execution in fillStroke function in class IntertwineAsLines and  class IntertwineAsPixelPerfect inside intertwines.h
- Added Stroke::erase function to Stroke class. It is needed inside IntertwineAsPixelPerfect class to get a clear m_pts (without the extra points due to mini L shapes, due to pixel perfect process). In fillStroke function in class  IntertwineAsPixelPerfect inside intertwiners.h, when it  executed, m_pts is delivered to polygon function instead of stroke argument in order to pass a pixel perfect processed vector instead of stroke vector which is a raw vector of vertices.