Minor fix, shape closing with Polygon Tool
The purpose of this fix is to match polygon preview with the polygon result when using Polygon Tool (before this fix, we were been see few pixels differences on the shape closing segment).
To achieve this, the way how we draw the entire joinStroke has to match how polygon function interpolate the final stroke which closes the shape.
So, the input vertices of doPointshapeLine function were swapped.